+++ packages/ckb-sdk-rpc/src/exceptions/index.ts
@@ -1,4 +1,4 @@
-export { HexStringWithout0xException } from '@nervosnetwork/ckb-sdk-utils/lib/exceptions'
-export * from './formatter'
-export * from './rpc'
-export * from './batch'
+export { HexStringWithout0xException } from '@nervosnetwork/ckb-sdk-utils'
+export * from './formatter.js'
+export * from './rpc.js'
+export * from './batch.js'