+++ packages/ckb-sdk-utils/src/exceptions/blake2b.ts
@@ -1,4 +1,4 @@
-import ErrorCode from './ErrorCode'
+import ErrorCode from './ErrorCode.js'
 
 export class OutLenTooSmallException extends Error {
   code = ErrorCode.ParameterInvalid