+++ packages/ckb-sdk-utils/src/exceptions/transaction.ts
@@ -1,4 +1,4 @@
-import ErrorCode from './ErrorCode'
+import ErrorCode from './ErrorCode.js'
 
 export class ReconciliationException extends Error {
   code = ErrorCode.ReconciliationFailed